+++ Editor/ChildComponents/ItemStyleDrawer.cs
@@ -24,6 +24,7 @@ namespace XCharts.Editor
                 PropertyField(prop, "m_CenterColor");
                 PropertyField(prop, "m_CenterGap");
                 PropertyField(prop, "m_BorderWidth");
+                PropertyField(prop, "m_BorderGap");
                 PropertyField(prop, "m_BorderColor");
                 PropertyField(prop, "m_BorderColor0");
                 PropertyField(prop, "m_BorderToColor");